requirements.txt
fastapi
uvicorn
python-dotenv
pdfplumber
python-docx
faiss-cpu
google-generativeai
requests


